Fix undefined offset notice first time the settings page is visited on a virgin site.
Fixes error notice:
`PHP Notice:  Undefined offset: 0 in /path/to/wp-content/plugins/tabify-edit-screen/features/detection/detection.php on line 76`

This error occurs when the settings page is visited before any posts have been added.
As this is likely a temporary situation, the time out for the transient is set to an hour to allow for users starting to add posts.